+++ CHANGELOG.rst
@@ -3,6 +3,15 @@ Changelog
 ==========
 
 +++++++++
+v1.4.0 (05/03/2024)
++++++++++
+
+**Added**
+
+- CTF endpoint for SPA now returns particle count for all image numbers
+- Standard deviation is returned for average ice thickness
+
++++++++++
 v1.3.2 (19/02/2024)
 +++++++++
 
@@ -10,6 +19,7 @@ v1.3.2 (19/02/2024)
 
 - Improved performance for :code:`/sessions` endpoint, only returns eBIC sessions
 - Fixed parameter generation for tomogram reprocessing
+- CTF endpoint for autoprocessing programs now returns particle count
 
 +++++++++
 v1.3.1 (14/02/2024)